Change expiresAt to uint64

+++ contracts/pandasia.sol
@@ -53,7 +53,7 @@ contract Pandasia is Ownable {
     uint256 balance; // current balance of asset in the airdrop
     bytes32 root; // optional merkle root for this airdrop
     uint256 claimAmount; // claimAmount claimable by each address
-    uint32 expiresAt; // time that airdop expires and no further claims can be made
+    uint64 expiresAt; // time that airdop expires and no further claims can be made
     bool onlyRegistered; // if onlyRegistered=true than addr must be in root AND merkleRoot, else an addr in root OR (previously seen valdiator in pandasia or googpool) is eligble
   }
 
@@ -61,7 +61,7 @@ contract Pandasia is Ownable {
   /*** Airdrop Functions                                                                                                              ***/
   /**************************************************************************************************************************************/
 
-  function newAirdrop(bytes32 root, bool onlyRegistered, address erc20, uint256 claimAmount, uint32 expiresAt) external returns (uint64) {
+  function newAirdrop(bytes32 root, bool onlyRegistered, address erc20, uint256 claimAmount, uint64 expiresAt) external returns (uint64) {
     if (erc20 == address(0)) {
       revert InvalidAddress();
     }
